% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/441_Package.R
\name{dividing}
\alias{dividing}
\title{Dividing Function}
\usage{
dividing(a, b)
}
\arguments{
\item{a}{this should be a float or an integer}

\item{b}{this should be a float or an integer}
}
\value{
Returns an integer or float that is the quotient of a and b
}
\description{
Dividing Function
}
\examples{
dividing(5,3)
}
