% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/441_Package.R
\name{mean}
\alias{mean}
\title{Mean Function}
\usage{
mean(x)
}
\arguments{
\item{x}{this should be a numeric vector}
}
\value{
Returns an integer or float that is the mean of numeric vector x
}
\description{
Mean Function
}
\examples{
x <- c(1,2,3,4,5,6,7,8,9)
mean(x)
}
