% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/441_Package.R
\name{multiplying}
\alias{multiplying}
\title{Multiplication Function}
\usage{
multiplying(a, b)
}
\arguments{
\item{a}{this should be a float or an integer}

\item{b}{this should be a float or an integer}
}
\value{
Returns an integer or float that is the product of a and b
}
\description{
Multiplication Function
}
\examples{
multiplying(5,3)
}
