% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/441_Package.R
\name{se}
\alias{se}
\title{Standard Error Function}
\usage{
se(x)
}
\arguments{
\item{x}{this should be a numeric vector}
}
\value{
Returns an integer or float that is the standard error of numeric vector x
}
\description{
Standard Error Function
}
\examples{
x <- c(1,2,3,4,5)
se(x)
}
