% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/441_Package.R
\name{subtracting}
\alias{subtracting}
\title{Subtraction Function
Title}
\usage{
subtracting(a, b)
}
\arguments{
\item{a}{this should be a float or an integer}

\item{b}{this should be a float or an integer}
}
\value{
Returns an integer or float that is the difference of a and b
}
\description{
Subtraction Function
Title
}
\examples{
subtracting(4,2)
}
